LPS-66316 Move Indexers to search package
