+++ modules/apps/mail/mail-service/src/main/java/com/liferay/mail/search/FolderIndexer.java
@@ -12,12 +12,12 @@
  * details.
  */
 
-package com.liferay.mail.util;
+package com.liferay.mail.search;
 
 import com.liferay.expando.kernel.model.ExpandoBridge;
 import com.liferay.expando.kernel.util.ExpandoBridgeIndexerUtil;
-import com.liferay.mail.model.Account;
-import com.liferay.mail.service.AccountLocalServiceUtil;
+import com.liferay.mail.model.Folder;
+import com.liferay.mail.service.FolderLocalServiceUtil;
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.IndexableActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
@@ -52,9 +52,9 @@ import org.osgi.service.component.annotations.Component;
  * @author Peter Fellwock
  */
 @Component(immediate = true, service = Indexer.class)
-public class AccountIndexer extends BaseIndexer<Account> {
+public class FolderIndexer extends BaseIndexer<Folder> {
 
-	public static final String CLASS_NAME = Account.class.getName();
+	public static final String CLASS_NAME = Folder.class.getName();
 
 	@Override
 	public String getClassName() {
@@ -62,10 +62,10 @@ public class AccountIndexer extends BaseIndexer<Account> {
 	}
 
 	@Override
-	protected void doDelete(Account account) throws Exception {
+	protected void doDelete(Folder folder) throws Exception {
 		SearchContext searchContext = new SearchContext();
 
-		searchContext.setCompanyId(account.getCompanyId());
+		searchContext.setCompanyId(folder.getCompanyId());
 		searchContext.setEnd(QueryUtil.ALL_POS);
 		searchContext.setSearchEngineId(getSearchEngineId());
 		searchContext.setSorts(SortFactoryUtil.getDefaultSorts());
@@ -75,7 +75,7 @@ public class AccountIndexer extends BaseIndexer<Account> {
 
 		booleanQuery.addRequiredTerm(Field.ENTRY_CLASS_NAME, CLASS_NAME);
 
-		booleanQuery.addRequiredTerm("accountId", account.getAccountId());
+		booleanQuery.addRequiredTerm("folderId", folder.getFolderId());
 
 		Hits hits = IndexSearcherHelperUtil.search(searchContext, booleanQuery);
 
@@ -88,18 +88,20 @@ public class AccountIndexer extends BaseIndexer<Account> {
 		}
 
 		IndexWriterHelperUtil.deleteDocuments(
-			getSearchEngineId(), account.getCompanyId(), uids,
+			getSearchEngineId(), folder.getCompanyId(), uids,
 			isCommitImmediately());
 	}
 
 	@Override
-	protected Document doGetDocument(Account account) throws Exception {
-		Document document = getBaseModelDocument(CLASS_NAME, account);
+	protected Document doGetDocument(Folder folder) throws Exception {
+		Document document = getBaseModelDocument(CLASS_NAME, folder);
 
-		ExpandoBridge expandoBridge = account.getExpandoBridge();
+		ExpandoBridge expandoBridge = folder.getExpandoBridge();
 
-		document.addKeyword("accountId", account.getAccountId());
-		document.addText(Field.NAME, account.getAddress());
+		document.addKeyword(Field.FOLDER_ID, folder.getFolderId());
+		document.addText(Field.NAME, folder.getDisplayName());
+
+		document.addKeyword("accountId", folder.getAccountId());
 
 		ExpandoBridgeIndexerUtil.addAttributes(document, expandoBridge);
 
@@ -115,19 +117,19 @@ public class AccountIndexer extends BaseIndexer<Account> {
 	}
 
 	@Override
-	protected void doReindex(Account account) throws Exception {
-		Document document = getDocument(account);
+	protected void doReindex(Folder folder) throws Exception {
+		Document document = getDocument(folder);
 
 		IndexWriterHelperUtil.updateDocument(
-			getSearchEngineId(), account.getCompanyId(), document,
+			getSearchEngineId(), folder.getCompanyId(), document,
 			isCommitImmediately());
 	}
 
 	@Override
 	protected void doReindex(String className, long classPK) throws Exception {
-		Account account = AccountLocalServiceUtil.getAccount(classPK);
+		Folder folder = FolderLocalServiceUtil.getFolder(classPK);
 
-		doReindex(account);
+		doReindex(folder);
 	}
 
 	@Override
@@ -139,26 +141,26 @@ public class AccountIndexer extends BaseIndexer<Account> {
 
 	protected void reindexMessages(long companyId) throws PortalException {
 		final IndexableActionableDynamicQuery indexableActionableDynamicQuery =
-			AccountLocalServiceUtil.getIndexableActionableDynamicQuery();
+			FolderLocalServiceUtil.getIndexableActionableDynamicQuery();
 
 		indexableActionableDynamicQuery.setCompanyId(companyId);
 		indexableActionableDynamicQuery.setPerformActionMethod(
-			new ActionableDynamicQuery.PerformActionMethod<Account>() {
+			new ActionableDynamicQuery.PerformActionMethod<Folder>() {
 
 				@Override
-				public void performAction(Account account)
+				public void performAction(Folder folder)
 					throws PortalException {
 
 					try {
-						Document document = getDocument(account);
+						Document document = getDocument(folder);
 
 						indexableActionableDynamicQuery.addDocuments(document);
 					}
 					catch (PortalException pe) {
 						if (_log.isWarnEnabled()) {
 							_log.warn(
-								"Unable to index account " +
-									account.getAccountId(),
+								"Unable to index folder " +
+									folder.getFolderId(),
 								pe);
 						}
 					}
@@ -170,6 +172,6 @@ public class AccountIndexer extends BaseIndexer<Account> {
 		indexableActionableDynamicQuery.performActions();
 	}
 
-	private static final Log _log = LogFactoryUtil.getLog(AccountIndexer.class);
+	private static final Log _log = LogFactoryUtil.getLog(FolderIndexer.class);
 
 }
